# internal/routers/api/v1/article.go
/**
 * @Title  article
 * @Description  文章处理
 * @Author  沈来
 * @Update  2020/8/3 16:17
 **/
package v1

import (
	"github.com/gin-gonic/gin"
	"myBlog/pkg/app"
	"myBlog/pkg/errcode"
)

type Article struct{

}

func NewArticle() Article{
	return Article{}
}

// @Summary  获取单篇文章
// @Produce  json
// @Param  id path int true "文章ID"
// @Param  name query string false "文章名称" maxlength(100)
// @Param  state query int false "状态" Enums(0, 1) default(1)
// @Param  page query int false "页码"
// @Param  page_size query int false "每页数量"
// @Success  200 {object} model.ArticleSwagger "成功"
// @Failure  400 {object} errcode.Error "请求错误"
// @Failure  500 {object} errcode.Error "内部错误"
// @Router  /api/v1/articles/{id} [get]
func (a Article) Get(c *gin.Context){
	app.NewResponse(c).ToErrorResponse(errcode.ServerError)
	return
}

// @Summary  获取多篇文章
// @Produce  json
// @Param  name query string false "文章名称" maxlength(100)
// @Param  state query int false "状态" Enums(0, 1) default(1)
// @Param  page query int false "页码"
// @Param  page_size query int false "每页数量"
// @Success  200 {object} model.ArticleSwagger "成功"
// @Failure  400 {object} errcode.Error "请求错误"
// @Failure  500 {object} errcode.Error "内部错误"
// @Router  /api/v1/articles [get]
func (a Article) List(c *gin.Context){

}

// @Summary  新增文章
// @Produce  json
// @Param  name body string true "文章名称" minlength(3) maxlength(100)
// @Param  state body int false "状态" Enums(0, 1) default(1)
// @Param  created_by body string false "创建者" minlength(3) maxlength(100)
// @Success  200 {object} model.Tag "成功"
// @Failure  400 {object} errcode.Error "请求错误"
// @Failure  500 {object} errcode.Error "内部错误"
// @Router  /api/v1/articles [post]
func (a Article) Create(c *gin.Context){

}

// @Summary  更新文章
// @Produce  json
// @Param  id path int true "文章ID"
// @Param  name body string false "文章名称" minlength(3) maxlength(100)
// @Param  state body int false "状态" Enums(0, 1) default(1)
// @Param  modified_by body string true "修改者" minlength(3) maxlength(100)
// @Success  200 {array} model.Tag "成功"
// @Failure  400 {object} errcode.Error "请求错误"
// @Failure  500 {object} errcode.Error "内部错误"
// @Router  /api/v1/articles/{id} [put]
func (a Article) Update(c *gin.Context){

}

// @Summary  删除文章
// @Produce  json
// @Param  id path int true "文章ID"
// @Success  200 {string} string "成功"
// @Failure  400 {object} errcode.Error "请求错误"
// @Failure  500 {object} errcode.Error "内部错误"
// @Router  /api/v1/articles/{id} [delete]
func (a Article) Delete(c *gin.Context){

}